test: since there is now an inheritance system, it would be good to use it in tests.
 It's need to create a parameterized abstract test class that will store the Task Manager and contain tests for its methods. Then inherit from it, substitute a specific implementation, and add tests if necessary.
